Extract protocol management to own class
