+++ unity/EditorPlugin/AfterUnity56/EntryPoint.cs
@@ -6,6 +6,8 @@ namespace JetBrains.Rider.Unity.Editor.AfterUnity56
 {
   // DO NOT CHANGE NAME OR NAMESPACE!
   // Accessed from the package via reflection
+  // This class is only InitializeOnLoad when the plugin is loaded by Unity from the Assets folder. When the package
+  // explicitly loads the plugin from the product install folder, it will execute this class constructor.
   [InitializeOnLoad, PublicAPI]
   public static class EntryPoint
   {
@@ -18,14 +20,19 @@ namespace JetBrains.Rider.Unity.Editor.AfterUnity56
       if (UnityUtils.IsInBatchModeAndNotInRiderTests)
         return;
 
-      PluginEntryPoint.OnModelInitialization += Initialization.OnModelInitializationHandler;
-      PluginEntryPoint.OnModelInitialization += Navigation.Initialization.OnModelInitializationHandler;
-      PluginEntryPoint.OnModelInitialization += Packages.Initialization.OnModelInitializationHandler;
+      // Make sure the PluginEntryPoint class constructor has been called. This used to happen implicitly when this
+      // class accessed fields of PluginEntryPoint. Refactorings mean these fields are no longer there, so let's be
+      // explicit about the dependency.
+      PluginEntryPoint.EnsureInitialised();
+
+      UnityEditorProtocol.OnModelInitialization += Initialization.OnModelInitializationHandler;
+      UnityEditorProtocol.OnModelInitialization += Navigation.Initialization.OnModelInitializationHandler;
+      UnityEditorProtocol.OnModelInitialization += Packages.Initialization.OnModelInitializationHandler;
       AppDomain.CurrentDomain.DomainUnload += (_, __) =>
       {
-        PluginEntryPoint.OnModelInitialization -= Initialization.OnModelInitializationHandler;
-        PluginEntryPoint.OnModelInitialization -= Navigation.Initialization.OnModelInitializationHandler;
-        PluginEntryPoint.OnModelInitialization -= Packages.Initialization.OnModelInitializationHandler;
+        UnityEditorProtocol.OnModelInitialization -= Initialization.OnModelInitializationHandler;
+        UnityEditorProtocol.OnModelInitialization -= Navigation.Initialization.OnModelInitializationHandler;
+        UnityEditorProtocol.OnModelInitialization -= Packages.Initialization.OnModelInitializationHandler;
       };
     }
   }
